requirements.txt
Pillow==6.1.0
psutil==5.6.3
pypiwin32==223
pystray==0.14.4
requests==2.22.0


